% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_se.r
\name{add_se}
\alias{add_se}
\title{Compute standard error for group-level effects in linear models}
\usage{
add_se(model, name_f, name_x = "Intercept")
}
\arguments{
\item{model}{A fitted model object}

\item{name_f}{A character, the name of the factor variables}

\item{name_x}{A character, the name of a continuous variable interacting with name_f, default is Intercept}
}
\value{
A data frame where the rows are the different level of the factor, and the columns contain the following values:

Coef_link : fitted coefficient value on the link scale

SE : standard error of the fitted coefficient on the link scale

Coef_resp : fitted coefficient on the response scale (applying the inverse of the link function)

LCI, UCI : lower and upper bound of the 95\% confidence interval of the fitted coefficient on the response scale
}
\description{
\code{add_se} compute the coefficients, standard errors and confidence intervals for different levels of factor variables used in linear models.
}
\details{
Add_se currently support models fitted via: \code{\link[stats]{lm}}, \code{\link[stats]{lm}}, \code{\link[lme4]{lmer}}, \code{\link[lme4]{glmer}}, \code{\link[glmmTMB]{glmmTMB}}

Note that name_x must be the name of continuous variable interacting with the factor, factor - factor interaction are not supported.
}
\examples{
data(iris)
m <- lm(Sepal.Length ~ Sepal.Width * Species, data = iris)
#get the regression lines with SE and confidence intervals for
#the different iris species
add_se(m, name_f = "Species", name_x = "Sepal.Width")

}
